backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.5
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
google-generativeai==0.8.3
zhipuai==2.1.5.20241203
faster-whisper==1.0.3
pdfplumber==0.11.4
pytesseract==0.3.13
langdetect==1.0.9
python-jose[cryptography]==3.3.0
Pillow==10.4.0
python-multipart==0.0.12
supabase
openai==1.45.0
reportlab>=4.0.0
pypdf==4.1.0
